requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
transformers==4.37.0
accelerate==0.25.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
qwen-vl-utils==0.0.5